requirements.txt
torch==2.4.1                 
transformers==4.45.1     
scikit-learn==1.5.2        
numpy==2.1.1               
nats-py==2.9.0         
python-dotenv==1.0.1
pytest==8.3.3     
